Moved expand_qg method to be called in fetcher instead of in query planner
